Move useSessionStartTime to BaseStepBuilder
